added more cart error handling

+++ public/js/stall.js
@@ -3,16 +3,32 @@ import { LS_KEYS } from "./const.js";
 // TODO: make it work for all buttons
 const addCartBtn = document.querySelector(".addToCartBtn");
 
+function ifItemInCart(cart, item) {
+    if (cart[item.stallId]) {
+        if (cart[item.stallId].find((i) => i.itemCode == item.itemCode))
+            return true;
+    }
+    return false;
+}
+
 async function addToCart() {
     // TODO: GET the menu item details by id. ie for each button in the data- there's an attribute to store the id of that menu item so can GET /menuItemByStallCodeAndItemCode
     const item = {
         stallId: "DDDDDDD1-DDDD-DDDD-DDDD-DDDDDDDDDDDD",
         itemCode: "M001",
-        // TODO: update the quantity
-        quantity: 2,
+        quantity: 1,
     };
 
+    // if item already exists in cart, don't add it
     let cart = JSON.parse(localStorage.getItem(LS_KEYS.cart) ?? "{}");
+    const isExist = ifItemInCart(cart, item);
+    if (isExist) {
+        alert(
+            "Cannot add to cart as item is already in cart. Go to cart and increase the quantity",
+        );
+        return;
+    }
+
     if (cart[item.stallId]) cart[item.stallId].push(item);
     else cart[item.stallId] = [item];
     localStorage.setItem(LS_KEYS.cart, JSON.stringify(cart));
@@ -20,6 +36,6 @@ async function addToCart() {
     alert("Added to cart!");
 }
 
-// TODO: once clicked, there should be "+" and "-"  buttons instead
-// load the addToCartBtn in via js
+// TODO: once clicked, there should be "+" and "-"  buttons instead -> update quantity
+// load the addToCartBtn in js
 addCartBtn.addEventListener("click", addToCart);
